infra/requirements.txt
aws-cdk-lib==2.115.0
constructs>=10.0.0,<11.0.0
cdk-ecr-deployment
cdk-nag==2.27.216
cdklabs.aws-data-solutions-framework>=1.0.0